feat: :art: sea crea archivo para no mezclar los tipos del dominio dentro de un servicio
