+++ frontend/auth-frontend/src/services/monthlyPaymentsService.ts
@@ -1,24 +1,5 @@
 import api from "./authService";
-
-export interface MonthlyPayment {
-  id: string;
-  userId: string;
-  year: number;
-  month: number;
-  amount: number;
-  description?: string;
-  paymentDate: string;
-  createdAt: string;
-  updatedAt: string;
-  createdBy?: string;
-  user?: { id: string; name: string; email: string; rut: string };
-}
-
-export interface PaymentSummary {
-  year: number;
-  total: number;
-  months: Array<{ month: number; amount: number; description?: string }>;
-}
+import { MonthlyPayment, PaymentSummary } from "../types/payment";
 
 class MonthlyPaymentsService {
   async createPayment(
